Authentication-with-OOP-PHP
In the realm of web development, ensuring secure user authentication is paramount to safeguarding sensitive information and protecting user accounts. Object-Oriented Programming (OOP) in PHP provides an effective and organized approach to implementing authentication systems.

One of the key advantages of OOP is the ability to encapsulate code into reusable and modular components, promoting maintainability and scalability. When it comes to authentication, this approach allows developers to create a robust and flexible system.